pyproject.toml
[project]
name = "facilty-feed-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Huzaifa Ahmed",email = "huzaifa.ahmed01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "asyncpg (>=0.30.0,<0.31.0)",
    "aioboto3 (>=14.1.0,<15.0.0)",
    "pylint (>=3.3.6,<4.0.0)",
    "autopep8 (>=2.3.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "asynctest (>=0.13.0,<0.14.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
